refactoring drag and drop
